Modify service and dao layers to return a ResponseEntity<Object> instead of just an Object. Returning an object is not sufficient as many REST api calls use HTTP status alone to indicate the success, or failure, of an operation (i.e. 204 No Content).
I've also added RestProxyResponseErrorHandler, a ResponseErrorHandler which is always used and does nothing, but it does highlight some important assumptions:
RestProxy's sole responsibility is to relay requests and responses.
It's up to the client to deal with responses, whether they are successes or errors.

RestProxyServiceImpl has also been modified to allow any content type and to only attach a body when the request has a content length greater than zero.